+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.33 20??-??-??
+ - 2018-12-05 Added possibility to display links in footer section of customer and public interface.
  - 2018-11-27 Fixed bug#[14246](https://bugs.otrs.org/show_bug.cgi?id=14246) - Attachments with 0 byte do not show up.
  - 2018-11-20 Fixed bug#[14237](https://bugs.otrs.org/show_bug.cgi?id=14237) - Sorting of state menu entries is different then order in state overview screen.
  - 2018-11-19 Fixed bug#[14148](https://bugs.otrs.org/show_bug.cgi?id=14148) - Superfluous filter boxes and "Add Group" form in Group Management screen.